scanNumber and id as Int. Test corrected: 7 tests PepSpectraMatchSpecs LoaderMzXMLSpecs LoaderMaxQuantSpecs LoaderMGFSpecs LoaderMzIdentSpecs ParseProteinsMatchesSpecs MatchesMongoDBServiceSpecs

+++ app/ch/isbsib/proteomics/mzviz/controllers/matches/PSMController.scala
@@ -190,7 +190,7 @@ object PSMController extends MatchController {
       Action.async { implicit request =>
         MatchMongoDBService().findAllBySearchIdAndSpectrumId(
           SearchId(searchId),
-          SpectrumUniqueId(spectrumId)
+          SpectrumUniqueId(spectrumId.toInt)
         )
           .map { case psms =>
           render {
